Move some files, update deps, and add minor checks in scripts

+++ local/test.sh
@@ -3,6 +3,8 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd "$DIR" 
 
 # jq is a dependency
+command -v jq >/dev/null 2>&1 || { echo >&2 "jq is required but it's not installed. Aborting."; exit 1; }
+
 rm -rf tmp.txt
 
 echo "Testing /health"
